# Conflicts:
#	CHANGELOG.adoc
#	c2c-cli/src/test/groovy/org/certificateservices/custom/c2c/cli/command/ViewCommandSpec.groovy
#	c2c-common/src/test/groovy/org/certificateservices/custom/c2x/etsits102941/v131/util/Etsi102941CTLHelperSpec.groovy

+++ src/main/java/org/certificateservices/custom/c2x/asn1/coer/COERTag.java
@@ -86,14 +86,14 @@ public class COERTag implements COEREncodable{
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		if (tagNumber < 63) 
 		{
-			baos.write(tagClass | (int) tagNumber);
+			baos.write(tagClass | (byte) tagNumber);
 		}
 		else
 		{
 			baos.write(tagClass | 0x3F);
 			if (tagNumber < 128)
 			{
-				baos.write((int) tagNumber);
+				baos.write((byte) tagNumber);
 			}
 			else
 			{
